refactor: typesをdomainにリネームして、gameをdomain内に移動
