+++ app/domain/game/draw.go
