+++ app/domain/tetromino.go
@@ -1,4 +1,4 @@
-package types
+package domain
 
 import (
 	"image/color"
